Make sure pages are mprotected PROT_EXEC for amd64 support

+++ gcl/o/gbc.c
@@ -134,7 +134,7 @@ struct apage {
   char apage_self[PAGESIZE];
 };
 
-int maxpage;
+long maxpage;
 
 object sSAnotify_gbcA;
